renamed model file for clarity, created foundation for ai controller
