+++ controller/local_controller.py
@@ -1,11 +1,11 @@
-from model.model import Game
+from controller.controller import Controller
+from model.game_model import Game
 from view.game_view import GameView
 
 
-class LocalController:
+class LocalController(Controller): # noqa
     def __init__(self, model: Game):
-        self.model = model
-        self.view = None
+        super.init(model)
 
     def run_game(self):
         game_ended = False
@@ -67,21 +67,6 @@ class LocalController:
             self.view.display_board()
             self.view.display_curr_player(self.model.curr_player)
 
-    def change_board_size(self, new_size):
-        if (new_size % 2) != 0 or new_size < 6:
-            return -1
-        if self.view is not None:
-            self.model.set_board_size(new_size)
-
-    def reset_game(self):
-        size = self.model.board.shape[0]
-        self.model = Game()
-        self.model.set_board_size(size)
-        # self.model.set_board_size(self.model.board.shape[0])
-
-    def set_view(self, view: GameView):
-        self.view = view
-
     def get_move(self):
         move = [0]
         valid = False
@@ -99,7 +84,3 @@ class LocalController:
                 move = [0]
                 continue
         return row, col
-
-    def get_leaderboard(self):
-
-        return
