+++ model/AiStrategy
@@ -1,0 +1,7 @@
+from abc import ABC, abstractmethod
+
+
+class AiStrategy(ABC):
+    @abstractmethod
+    def determine_move(self, board):
+        pass
