+++ model/EasyAi
@@ -1,0 +1,7 @@
+from model import AiStrategy
+
+
+class EasyAi(AiStrategy):
+    def determine_move(self, board):
+        pass
+
